+++ application/views/homepage.php
@@ -1,5 +1,7 @@
 <div class="row">
-    {authors}
-    <div class="span4"><img src="/pix/{who}.jpg" title="{who} plays {role}"/></div>
-    {/authors}
+        <div class="span4">
+                <a href="/welcome/show/{key}">
+                        <img src="/pix/{who}.jpg" title="{who} plays {role}"/>
+                </a>
+        </div>
 </div>